Add files via upload
Problem statement:
 Given a dataset on air quality, build a machine learning model 
based on support vector regression and random forest regression, which should be able 
to predict the temperature of the day based on air concentration.

Background:
 Air quality is a measure of how clean or polluted the air is. Monitoring air 
quality is essential because polluted air can be detrimental to our wellbeing and the safety 
of the world. Air quality depends on many of its components such as CO2, NOx etc. 

About the Dataset:
 The dataset given contains 9358 instances and 15 attributes. It 
includes measurements of air composition such as NOx, NO2, CO etc. The datasets can 
be obtained from the following Machine Learning Repository with the following 
URL:
https://archive.ics.uci.edu/ml/datasets/Air+quality